warnings batch 1 — real correctness fixes + drop spurious PACKED on FAMs
Six concrete fixes plus a header-wide cleanup. The build was emitting
~1024 warnings on a fresh GCC; this batch takes 388 of them off the
table.

src/xfers.c:535 (-Warray-bounds, real bug)
  memset(&fi, 0, sizeof(struct hfs_cap_info)) — but fi is
  struct hfsinfo, which is 76 bytes smaller than hfs_cap_info.
  We were zeroing 76 bytes past the end of fi on every preview
  download. Switch to sizeof(fi).

src/chat.c:789 (-Wformat-overflow + -Wrestrict, real bug)
  sprintf(nick_buf, "%s%s ", nick_buf, nick) — sprintf with
  source and destination overlapping is undefined behaviour, and
  with enough match candidates would also have walked off the
  end of the 2048-byte buffer. Replace with a snprintf+offset
  loop that bounds-checks every append.

src/hfs.c:373,378 (-Wstringop-overflow)
  read(f, &fi->create_time, 8) into a 4-byte field, and
  read(f, fi->type, 8) into a 4-byte array — both reading 8
  bytes through a 4-byte pointer. Intent was to read two
  adjacent struct fields in a single syscall (DATES is
  create_time+modify_time, FINFO is type+creator). Read into a
  local 8-byte buffer and memcpy into the two fields explicitly;
  same wire result, no out-of-bounds writes through a typed
  pointer. Also fixes the dual bug in the FINFO error path
  (memset(fi->type, 0, 8) was writing past type into creator).

src/connect.c:343,349,355,822,827,832 (-Wformat)
  fprintf("%c", len) where len is size_t. Worked accidentally
  on x86 because the pushed value's low byte happened to land
  where %c looked for an int. Cast to (int) explicitly.

src/hotline.h, src/files.c:1330 (-Wattributes ×317)
  PACKED on a flexible-array-member or a guint8[] field is
  ignored by GCC (you can't pack an alignment-1 array any
  further). Drop the redundant PACKED on the four hotline.h
  FAMs and the files.c x_fhdr name[1]. Other PACKED-ed fields
  in those structs stay — only the byte-array tail loses it.

+++ src/connect.c
@@ -340,19 +340,19 @@ convert_bookmark (AdwAlertDialog *dialog, const char *response, gpointer data)
 
 	len = strlen (login);
 	len_total = 33 - len;
-	fprintf (bm, "%c", len);
+	fprintf (bm, "%c", (int) len);
 	fprintf (bm, "%s", login);
 	fwrite (zeros, 1, len_total, bm);
 
 	len = strlen (pass);
 	len_total = 33 - len;
-	fprintf (bm, "%c", len);
+	fprintf (bm, "%c", (int) len);
 	fprintf (bm, "%s", pass);
 	fwrite (zeros, 1, len_total, bm);
 
 	len = strlen (server);
 	len_total = 256 - len;
-	fprintf (bm, "%c", len);
+	fprintf (bm, "%c", (int) len);
 	fprintf (bm, "%s", server);
 
 	/* secure:0, compress:0, cipher:0 */
@@ -819,17 +819,17 @@ bookmark_save_response (AdwAlertDialog *dialog, const char *response, gpointer d
 
 	len = strlen (login);
 	len_total = 33 - len;
-	fprintf (bookmark, "%c%s", len, login);
+	fprintf (bookmark, "%c%s", (int) len, login);
 	fwrite (zeros, 1, len_total, bookmark);
 
 	len = strlen (pass);
 	len_total = 33 - len;
-	fprintf (bookmark, "%c%s", len, pass);
+	fprintf (bookmark, "%c%s", (int) len, pass);
 	fwrite (zeros, 1, len_total, bookmark);
 
 	len = strlen (server_str);
 	len_total = 256 - len;
-	fprintf (bookmark, "%c%s", len, server_str);
+	fprintf (bookmark, "%c%s", (int) len, server_str);
 
 	fprintf (bookmark, "%c%c%c", secure, compress, cipher);
 	len_total -= 3;
